Move LineSampler into hephaestus library and add gslib as dependency. Also switch to using ninja as build generator
