+++ CMakeLists.txt
@@ -2,20 +2,22 @@ cmake_minimum_required(VERSION 3.8)
 set(CMAKE_POLICY_DEFAULT_CMP0077 NEW)
 include(FetchContent)
 
-# no space in these names! used as shared object name and python module name
+# Set name of project and library
 set(PROJECT_NAME hephaestus)
 set(PROJECT_LIB_NAME ${PROJECT_NAME}-lib)
 
 project(${PROJECT_NAME} VERSION 0.0.0 LANGUAGES CXX)
-find_package(Boost REQUIRED COMPONENTS ${Boost_PACKAGES})
 
-# Import MFEM. The following variables can be used to help CMake find MFEM:
+############################################################
+# Find and import MFEM 
+############################################################
+# The following variables can be used to help CMake find MFEM:
 #  * MFEM_DIR - absolute path to the MFEM build or install prefix.
 #  * mfem_DIR - absolute path to where MFEMConfig.cmake is.
-message(STATUS "Looking for mfem ...")
+message(STATUS "Looking for MFEM ...")
 
 set(MFEM_DIR "${PROJECT_SOURCE_DIR}/../mfem/build" CACHE PATH "Path to the MFEM build or install prefix.")
-set(MFEM_COMMON_INCLUDES "${MFEM_DIR}/../mfem/miniapps/common" CACHE PATH "Path to the MFEM common miniapp headers.")
+set(MFEM_COMMON_INCLUDES "${MFEM_DIR}/../miniapps/common" CACHE PATH "Path to the MFEM common miniapp headers.")
 
 if (MFEM_DIR)
    find_package(mfem REQUIRED NAMES MFEM HINTS "${MFEM_DIR}"
@@ -27,18 +29,18 @@ endif()
 ############################################################
 # Check C++ compiler is the same as used in MFEM 
 ############################################################
-message(STATUS "Found mfem config in: ${mfem_DIR} (version ${MFEM_VERSION})")
+message(STATUS "Found MFEM config in: ${mfem_DIR} (version ${MFEM_VERSION})")
 
 # Use the same C++ compiler as MFEM. This is needed when MFEM was built using
 # an MPI wrapper and we do not have explicitly the MPI compile and link flags.
-message(STATUS "Mfem compiler is: ${MFEM_CXX_COMPILER}")
+message(STATUS "MFEM compiler is: ${MFEM_CXX_COMPILER}")
 
 if (NOT CMAKE_CXX_COMPILER AND MFEM_CXX_COMPILER)
   set(CMAKE_CXX_COMPILER "${MFEM_CXX_COMPILER}")
 endif()
 
 message(STATUS "MFEM_INCLUDE_DIRS is set to: ${MFEM_INCLUDE_DIRS}")
-message(STATUS "MFEM_LIBRARIES is set to: ${MFEM_LIBRARIES}")
+message(STATUS "MFEM_COMMON_INCLUDES is set to: ${MFEM_COMMON_INCLUDES}")
 
 find_library(MFEM_COMMON_LIBRARY mfem-common HINTS ${mfem_DIR}/miniapps/common)
 message(STATUS "MFEM_COMMON_LIBRARY is found in the folder: ${MFEM_COMMON_LIBRARY}")
@@ -49,7 +51,8 @@ enable_testing()
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_SOURCE_DIR}/bin)
 ## put all libraries in lib
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${PROJECT_SOURCE_DIR}/lib)
-
+## export compile commands for clang-format
+set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 ############################################################
 # Fetch spdlog for logging purposes
@@ -85,9 +88,7 @@ elseif(BUILD_TYPE STREQUAL "RELEASE")
   set(TEST_LIBRARIES "")
 endif()
 
-############################################################
 # Check clang is being used if build type is set to "DEBUG-MEMORY"
-############################################################
 if(BUILD_TYPE STREQUAL "DEBUG-MEMORY" AND NOT ${CMAKE_CXX_COMPILER} STREQUAL "/usr/bin/clang++")
     message(FATAL_ERROR "DEBUG-MEMORY option requires the clang++ compiler.")
 endif()
