+++ src/auxsolvers/auxsolvers.hpp
@@ -9,6 +9,8 @@
 #include "vector_coefficient_aux.hpp"
 #include "vector_gridfunction_cross_product_aux.hpp"
 #include "vector_gridfunction_dot_product_aux.hpp"
+#include "line_sampler_aux.hpp"
+#include "flux_monitor_aux.hpp"
 
 // Specify classes that perform auxiliary calculations on GridFunctions or
 // Coefficients.
